refactor: controller 로직을 분리합니다.
