+++ src/backend/auth-server/src/controllers/index.ts
@@ -1,0 +1,1 @@
+export { default as authController } from './auth';
